# .docs/.openapi/api.yaml
openapi: 3.0.3
info:
  contact:
    email: andreas.rauber@tuwien.ac.at
    name: Prof. Andreas Rauber
  description: >
    The merged REST API of DBRepo for users, developers and data stewards to be
    accessed publicly. Have a look at

    the [source
    code](https://gitlab.phaidra.org/fair-data-austria-db-repository/fda-services)
    for non-public endpoints

    that are used between the services themselves.
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0'
  title: DBRepo REST API
  version: 1.6.1
servers:
  - description: Test Instance
    url: 'https://test.dbrepo.tuwien.ac.at'
  - description: Local Instance
    url: 'http://localhost'
...
paths:
...
  '/api/database/{databaseId}/access/{userId}':
    get:
      tags:
        - access-endpoint
      summary: Find/Check access
      description: >-
        Finds or checks access of a user with given id to a database with given
        id. Requests with HTTP method **GET** return the access object, requests
        with HTTP method **HEAD** only the status. When the user has at least
        *READ* access, the status 200 is returned, 403 otherwise. Requires role
        `check-database-access` or `check-foreign-database-access`.
      operationId: find
      parameters:
        - name: databaseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Found database access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseAccessDto'
        '403':
          description: No access to this database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '404':
          description: Database not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      security:
        - bearerAuth: []
        - basicAuth: []
    put:
      tags:
        - access-endpoint
      summary: Modify access
      description: >-
        Modifies access of a user with given id to database with given id.
        Requires role `update-database-access`.
      operationId: update_5
      parameters:
        - name: databaseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccessDto'
        required: true
      responses:
        '202':
          description: Modified access
        '400':
          description: Modify access query or database connection is malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '403':
          description: >-
            Modify access not permitted when no access is granted in the first
            place
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '404':
          description: Database or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '502':
          description: >-
            Access could not be updated due to connection error in the data
            service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '503':
          description: Access could not be updated in the data service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      security:
        - bearerAuth: []
        - basicAuth: []
    post:
      tags:
        - access-endpoint
      summary: Give access
      description: >-
        Give a user with given id access to some database with given id.
        Requires role `create-database-access`.
      operationId: create_8
      parameters:
        - name: databaseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccessDto'
        required: true
      responses:
        '202':
          description: Granting access succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseAccessDto'
        '400':
          description: Granting access query or database connection is malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '403':
          description: Failed giving access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '404':
          description: Database or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '502':
          description: Access could not be created due to connection error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '503':
          description: Access could not be created in the data service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      security:
        - bearerAuth: []
        - basicAuth: []
    delete:
      tags:
        - access-endpoint
      summary: Delete access
      description: >-
        Delete access of a user with id to a database with id. Requires role
        `delete-database-access`.
      operationId: revoke
      parameters:
        - name: databaseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: Deleted access
        '400':
          description: Modify access query or database connection is malformed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '403':
          description: Revoke of access not permitted as no access was found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '404':
          description: 'User, database with access was not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '502':
          description: Access could not be created due to connection error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '503':
          description: Access could not be revoked in the data service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      security:
        - bearerAuth: []
        - basicAuth: []
    head:
      tags:
        - access-endpoint
      summary: Find/Check access
      description: >-
        Finds or checks access of a user with given id to a database with given
        id. Requests with HTTP method **GET** return the access object, requests
        with HTTP method **HEAD** only the status. When the user has at least
        *READ* access, the status 200 is returned, 403 otherwise. Requires role
        `check-database-access` or `check-foreign-database-access`.
      operationId: find_1
      parameters:
        - name: databaseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Found database access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatabaseAccessDto'
        '403':
          description: No access to this database
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
        '404':
          description: Database not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
      security:
        - bearerAuth: []
        - basicAuth: []
...
components:
  securitySchemes:
    basicAuth:
      in: header
      scheme: basic
      type: http
    bearerAuth:
      bearerFormat: JWT
      in: header
      scheme: bearer
      type: http
  schemas:
...
    ApiErrorDto:
      required:
        - code
        - message
        - status
      type: object
      properties:
        status:
          type: string
          example: NOT_FOUND
          enum:
            - 100 CONTINUE
            - 101 SWITCHING_PROTOCOLS
            - 102 PROCESSING
            - 103 EARLY_HINTS
            - 103 CHECKPOINT
            - 200 OK
            - 201 CREATED
            - 202 ACCEPTED
            - 203 NON_AUTHORITATIVE_INFORMATION
            - 204 NO_CONTENT
            - 205 RESET_CONTENT
            - 206 PARTIAL_CONTENT
            - 207 MULTI_STATUS
            - 208 ALREADY_REPORTED
            - 226 IM_USED
            - 300 MULTIPLE_CHOICES
            - 301 MOVED_PERMANENTLY
            - 302 FOUND
            - 302 MOVED_TEMPORARILY
            - 303 SEE_OTHER
            - 304 NOT_MODIFIED
            - 305 USE_PROXY
            - 307 TEMPORARY_REDIRECT
            - 308 PERMANENT_REDIRECT
            - 400 BAD_REQUEST
            - 401 UNAUTHORIZED
            - 402 PAYMENT_REQUIRED
            - 403 FORBIDDEN
            - 404 NOT_FOUND
            - 405 METHOD_NOT_ALLOWED
            - 406 NOT_ACCEPTABLE
            - 407 PROXY_AUTHENTICATION_REQUIRED
            - 408 REQUEST_TIMEOUT
            - 409 CONFLICT
            - 410 GONE
            - 411 LENGTH_REQUIRED
            - 412 PRECONDITION_FAILED
            - 413 PAYLOAD_TOO_LARGE
            - 413 REQUEST_ENTITY_TOO_LARGE
            - 414 URI_TOO_LONG
            - 414 REQUEST_URI_TOO_LONG
            - 415 UNSUPPORTED_MEDIA_TYPE
            - 416 REQUESTED_RANGE_NOT_SATISFIABLE
            - 417 EXPECTATION_FAILED
            - 418 I_AM_A_TEAPOT
            - 419 INSUFFICIENT_SPACE_ON_RESOURCE
            - 420 METHOD_FAILURE
            - 421 DESTINATION_LOCKED
            - 422 UNPROCESSABLE_ENTITY
            - 423 LOCKED
            - 424 FAILED_DEPENDENCY
            - 425 TOO_EARLY
            - 426 UPGRADE_REQUIRED
            - 428 PRECONDITION_REQUIRED
            - 429 TOO_MANY_REQUESTS
            - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
            - 451 UNAVAILABLE_FOR_LEGAL_REASONS
            - 500 INTERNAL_SERVER_ERROR
            - 501 NOT_IMPLEMENTED
            - 502 BAD_GATEWAY
            - 503 SERVICE_UNAVAILABLE
            - 504 GATEWAY_TIMEOUT
            - 505 HTTP_VERSION_NOT_SUPPORTED
            - 506 VARIANT_ALSO_NEGOTIATES
            - 507 INSUFFICIENT_STORAGE
            - 508 LOOP_DETECTED
            - 509 BANDWIDTH_LIMIT_EXCEEDED
            - 510 NOT_EXTENDED
            - 511 NETWORK_AUTHENTICATION_REQUIRED
        message:
          type: string
          example: Error message
        code:
          type: string
          example: error.service.code
...
    UserBriefDto:
      required:
        - id
        - username
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 1ffc7b0e-9aeb-4e8b-b8f1-68f3936155b4
        username:
          type: string
          description: Only contains lowercase characters
          example: jcarberry
        name:
          type: string
          example: Josiah Carberry
        orcid:
          type: string
          example: 0000-0002-1825-0097
        qualified_name:
          type: string
          example: Josiah Carberry — @jcarberry
        given_name:
          type: string
          example: Josiah
        family_name:
          type: string
          example: Carberry
...
    DatabaseAccessDto:
      required:
        - type
        - user
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserBriefDto'
        type:
          type: string
          example: read
          enum:
            - read
            - write_own
            - write_all
...
    CreateAccessDto:
      required:
        - type
      type: object
      properties:
        type:
          type: string
          example: read
          enum:
            - read
            - write_own
            - write_all
...